+++ packages/gatsby-source-contentful/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.8.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-contentful@5.7.0-next.2...gatsby-source-contentful@5.8.0-next.0) (2021-06-03)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+# [5.7.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-contentful@5.7.0-next.1...gatsby-source-contentful@5.7.0-next.2) (2021-06-02)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
 # [5.7.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-contentful@5.7.0-next.0...gatsby-source-contentful@5.7.0-next.1) (2021-05-31)
 
 ### Bug Fixes